+++ csense-kotlin-jvm/src/main/kotlin/csense/kotlin/extensions/coroutines/CoroutineScope.kt
@@ -1,0 +1,14 @@
+package csense.kotlin.extensions.coroutines
+
+import csense.kotlin.CoroutineScopeFunction0
+import kotlinx.coroutines.*
+
+
+fun <T> CoroutineScope.asyncIO(
+        action: CoroutineScopeFunction0<T>
+): Deferred<T> = async(Dispatchers.IO, block = action)
+
+fun CoroutineScope.launchIO(
+        action: CoroutineScopeFunction0<Unit>
+): Job = launch(Dispatchers.IO, block = action)
+